fix(user): 用户 ID 消费链路统一为 Long，修复查看他人用户页 user_id=0 导致请求错误

+++ app/src/main/java/ceui/lisa/core/BaseRepo.java
@@ -9,7 +9,7 @@ public class BaseRepo {
         Common.showLog("BaseRepo " + getClass().getSimpleName() + " newInstance");
     }
 
-    public int currentUserID() {
-        return (int) SessionManager.INSTANCE.getLoggedInUid();
+    public long currentUserID() {
+        return SessionManager.INSTANCE.getLoggedInUid();
     }
 }
