+++ app/src/main/java/ceui/lisa/database/SearchDao.kt
@@ -77,20 +77,20 @@ interface SearchDao {
     fun getMutedUser(limit: Int, offset: Int): List<MuteEntity>
 
     @Query("SELECT * FROM tag_mute_table WHERE type = 3 AND id = :userID LIMIT 1")
-    fun getUserMuteEntityByID(userID: Int): MuteEntity?
+    fun getUserMuteEntityByID(userID: Long): MuteEntity?
 
     // 标签屏蔽记录的 id 就是 tag 名的 hashCode，见 PixivOperate.muteTag
     @Query("SELECT * FROM tag_mute_table WHERE type = 0 AND id = :tagId LIMIT 1")
     fun getTagMuteEntityByID(tagId: Int): MuteEntity?
 
     @Query("SELECT * FROM tag_mute_table WHERE type = 3 AND id = :userID LIMIT 1")
-    fun getUserMuteEntityByIDLiveData(userID: Int): LiveData<MuteEntity>
+    fun getUserMuteEntityByIDLiveData(userID: Long): LiveData<MuteEntity>
 
     @Query("SELECT * FROM tag_mute_table WHERE type = 1 AND id = :illustId LIMIT 1")
     fun getIllustMuteEntityByID(illustId: Int): LiveData<MuteEntity>
 
     @Query("SELECT * FROM tag_mute_table WHERE type = 4 AND id = :userID LIMIT 1")
-    fun getBlockMuteEntityByID(userID: Int): MuteEntity?
+    fun getBlockMuteEntityByID(userID: Long): MuteEntity?
 
     @Delete
     fun deleteMuteEntity(muteEntity: MuteEntity)
