+++ models/src/main/java/ceui/lisa/models/UserContainer.java
@@ -2,5 +2,5 @@ package ceui.lisa.models;
 
 public interface UserContainer {
 
-    int getUserId();
+    long getUserId();
 }
